XCode 9 / Swift 4 / iOS 11 work starts
The feature/ios10 branch never needed to go anywhere as the application worked fine. In this release, based on beta versions of XCode and iOS, new compiler warnings, Swift 4 upgrades and things like a revisited set of autolayout constraints are all used to improve performance and solve bugs that showed up on iOS 11 beta.