Swift 4; constraint fixes; toolbar height fixes.
Keyboard accessory views on the iPhone differ in height between portrait and landscape mode. For the "enter stop by ID" view, starting in portrait and rotating to landscape would result in a slight overlap of keyboard and text fields because the height change isn't reflected. After attempting a large number of solutions, the only thing that worked was to resign first responder then re-acquire it before and after rotation, causing the keyboard to show and hide. This would appear to be an iOS limitation (arguably bug) and the workaround is a little ugly, but less jarring than the incorrect toolbar heights on iPhone occurring after rotations.

For consistency, the same behaviour is implemented for the "edit description" view too, although it does not strictly need it. There is sufficient difference between the two to just copy and modify code for now, but if a third use case arises then this should get split out into a common base class that's flexible enough to accommodate one or many input fields and toolbars.