XCode signing issue resolution.
A prior XCode's automatic signing had assigned a different identifier to the WatchKit extensions. Upon submission of the previously tagged 1.6, iTunes Connect warned about the mismatch. After playing around a while, I realised it was best to let XCode (9 beta) work on fully automatic settings, assigning the same identifier to the WatchKit components as the parent application; reading Apple's documentation, it doesn't seem like it will be using any features that should be broken by the change of ID.

To enforce a new binary upload for double-checking, I've had to bump the version number and, since I lost confidence in the ability of this to run on iOS 8 and it contains nothing useful for anyone other than iOS 11 users strictly speaking, I've moved the minimum version up to 9.3 in passing. Will re-tag shortly.